Cleaned up references (except for in the notebooks), changed some more structural things
