+++ madminer/plotting.py
