+++ madminer/sampling.py
@@ -5,11 +5,12 @@ import numpy as np
 import collections
 import six
 
-from madminer.utils.interfaces.hdf5 import load_madminer_settings, madminer_event_loader, save_preformatted_events_to_madminer_file
+from madminer.utils.interfaces.hdf5 import load_madminer_settings, madminer_event_loader
+from madminer.utils.interfaces.hdf5 import save_preformatted_events_to_madminer_file
 from madminer.utils.analysis import get_theta_value, get_theta_benchmark_matrix, get_dtheta_benchmark_matrix
 from madminer.utils.analysis import extract_augmented_data, parse_theta
 from madminer.morphing import AdvancedMorpher
-from madminer.utils.utils import general_init, format_benchmark, create_missing_folders, shuffle, balance_thetas
+from madminer.utils.various import general_init, format_benchmark, create_missing_folders, shuffle, balance_thetas
 
 
 def combine_and_shuffle(input_filenames,
