+++ madminer/utils/various.py
@@ -116,3 +116,29 @@ def balance_thetas(theta_sets_types, theta_sets_values):
             theta_sets_values[i] = [values[i % n_sets_before] for i in range(n_sets)]
 
     return theta_sets_types, theta_sets_values
+
+
+
+
+def load_and_check(filename, warning_threshold=1.e9):
+    if filename is None:
+        return None
+
+    data = np.load(filename)
+
+    n_nans = np.sum(np.isnan(data))
+    n_infs = np.sum(np.isinf(data))
+    n_finite = np.sum(np.isfinite(data))
+
+    if n_nans + n_infs > 0:
+        logging.warning('Warning: file %s contains %s NaNs and %s Infs, compared to %s finite numbers!',
+                        filename, n_nans, n_infs, n_finite)
+
+    smallest = np.nanmin(data)
+    largest = np.nanmax(data)
+
+    if np.abs(smallest) > warning_threshold or np.abs(largest) > warning_threshold:
+        logging.warning('Warning: file %s has some large numbers, rangin from %s to %s',
+                        filename, smallest, largest)
+
+    return data
